Add Data upload functionality so that users can upload a csv and that data can be parsed as sensor data. Fixes #220
